functions/requirements.txt
firebase_functions
openai
python-dotenv